Refactoring reducers to own area, plus use of combineReducers
