+++ _assets/javascripts/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux'
+import answer from './answer'
+
+export default combineReducers({
+  answer
+})
